create an ArrayList of Asset objects.
Load it with your Assets. Include at least 2 houses (you have a vacation home!) and at least two vehicles. Use descriptions like "my house" or "Tom's truck" for the assets.

Now, loop thru the Asset collection displaying the description of each asset, the date you acquired it, how much you paid for it, and its value.

When that works, go back and include in the display either the address of the house or the year and make/model of the vehicle. You will need to use instanceof when you loop through the assets to detect the type of asset it is. Once you know it is a House or Vehicle, you will need to downcast it so that you can call the methods of the specific type.