Exercise 4-1
Create a copy of the Assets project and name it AbstractAssets. You will work in the AbstractAssets project.

Part 1 Modify your Asset class to be an abstract class. Then make the getValue() method an abstract method.

Re-test your application.

Try to create an instance of an Asset.

Part 2 Create a new class that extends Asset, the Cash class. This represents the cash that you stash away under your mattresses.